requirements.txt
Django>=4.2
djangorestframework>=3.14
drf-spectacular>=0.27
django-filter>=23.0
Pillow>=10.0
